# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "银行外包供应商风险 Multi-Agent 服务",
    "description": "当前风险分级 + 四智能体协同分析 + 结构化风险报告（不含预测）",
    "version": "0.3.0"
  },
  "paths": {
...
    "/agent/analyze": {
      "post": {
        "summary": "单/批量风险分析",
        "description": "批量分析入口：单供应商传一个 id，批量传多个或留空。\n\n``async`` 为 true 且为批量模式时，立即返回 ``task_id``，\n通过 ``GET /agent/task/{task_id}`` 轮询结果。",
        "operationId": "agent_analyze_agent_analyze_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Agent Analyze Agent Analyze Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/task/{task_id}": {
      "get": {
        "summary": "查询异步分析任务",
        "operationId": "agent_task_agent_task__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Agent Task Agent Task  Task Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/report/{supplier_id}": {
      "get": {
        "summary": "查询单个供应商结构化报告",
        "operationId": "agent_report_agent_report__supplier_id__get",
        "parameters": [
          {
            "name": "supplier_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Supplier Id"
            }
          },
          {
            "name": "current_week",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Current Week"
            }
          },
          {
            "name": "window_unit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "week",
                    "month"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Window Unit"
            }
          },
          {
            "name": "window_size",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 52,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Window Size"
            }
          },
          {
            "name": "audience",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "LEADERSHIP",
                "MANAGER"
              ],
              "type": "string",
              "default": "LEADERSHIP",
              "title": "Audience"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Agent Report Agent Report  Supplier Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/tools/supplier/{supplier_id}/evidence": {
      "get": {
        "summary": "证据取数（便于替换为后端真实接口）",
        "operationId": "supplier_evidence_agent_tools_supplier__supplier_id__evidence_get",
        "parameters": [
          {
            "name": "supplier_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Supplier Id"
            }
          },
          {
            "name": "current_week",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Current Week"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Supplier Evidence Agent Tools Supplier  Supplier Id  Evidence Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/tools/supplier/{supplier_id}/history": {
      "get": {
        "summary": "历史风险概况（分级 + 趋势 + 维度）",
        "operationId": "supplier_history_agent_tools_supplier__supplier_id__history_get",
        "parameters": [
          {
            "name": "supplier_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Supplier Id"
            }
          },
          {
            "name": "current_week",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Current Week"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Supplier History Agent Tools Supplier  Supplier Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/review": {
      "post": {
        "summary": "人工审核回写",
        "description": "把人工审核结论写回审核记录；同一 run_id 不允许重复提交。",
        "operationId": "submit_review_agent_review_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Submit Review Agent Review Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/redeclare": {
      "post": {
        "summary": "供应商重新申报（预留端点）",
        "description": "供应商侧重新申报受理（预留）。\n\n说明：是否**允许**重新申报由确定性处置门禁在分析与人工复核阶段判定\n（报告 ``disposition.can_redeclare``）。本端点只做登记：把供应商补充材料\n写入人工复核队列（``review_result=REDECLARED``），交由后端通知与人工复核处理。\n是否落在「问题较小」分支需先由后端查询该供应商报告确认。",
        "operationId": "submit_redeclare_agent_redeclare_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedeclareRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Submit Redeclare Agent Redeclare Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/agent/reviews": {
      "get": {
        "summary": "查询已回写的人工审核记录",
        "operationId": "list_reviews_agent_reviews_get",
        "parameters": [
          {
            "name": "supplier_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Supplier Id"
            }
          },
          {
            "name": "disposition_state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "按处置状态过滤",
              "title": "Disposition State"
            },
            "description": "按处置状态过滤"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response List Reviews Agent Reviews Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AnalyzeRequest": {
        "properties": {
          "supplier_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Supplier Ids",
            "description": "留空则分析全部供应商"
          },
          "current_week": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 52.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Week",
            "description": "留空则取数据中最大周次"
          },
          "enable_live_llm": {
            "type": "boolean",
            "title": "Enable Live Llm",
            "description": "默认尝试真实大模型；不可用时自动回退确定性文案",
            "default": true
          },
          "include_reports": {
            "type": "boolean",
            "title": "Include Reports",
            "description": "是否在响应中返回全部报告（同步批量模式生效）",
            "default": true
          },
          "watchlist_size": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Watchlist Size",
            "default": 10
          },
          "async": {
            "type": "boolean",
            "title": "Async",
            "description": "true 时批量分析转异步，立即返回 task_id",
            "default": false
          },
          "window_unit": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "week",
                  "month"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Unit",
            "description": "统计窗口单位：week（按周）/ month（按月，1月=4周）；留空取配置默认"
          },
          "window_size": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 52.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Size",
            "description": "窗口大小（周数或月数）"
          },
          "audience": {
            "type": "string",
            "enum": [
              "LEADERSHIP",
              "MANAGER"
            ],
            "title": "Audience",
            "description": "请求方角色：LEADERSHIP=高级管理人员；MANAGER=普通管理人员",
            "default": "LEADERSHIP"
          }
        },
        "type": "object",
        "title": "AnalyzeRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RedeclareRequest": {
        "properties": {
          "supplier_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Supplier Id"
          },
          "redeclare_reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Redeclare Reason"
          },
          "attachment_refs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Attachment Refs"
          },
          "submitted_by": {
            "type": "string",
            "maxLength": 100,
            "title": "Submitted By",
            "default": ""
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "supplier_id",
          "redeclare_reason"
        ],
        "title": "RedeclareRequest",
        "description": "供应商重新申报请求（预留端点，智能体只做登记与受理）。"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}